fix: resolve circular dependency in AI consensus service
- Move shared types to new ai_consensus_types.py file
- Update imports in ai_consensus.py and ai_consensus_core.py
- Fix circular dependency between consensus modules
